+++ install/requirements-conda.txt
@@ -1,0 +1,1 @@
+pythonocc
