+++ install/requirements-pip.txt
@@ -1,0 +1,3 @@
+PySide6
+ujson
+pyqtribbon
