+++ plugins/sketcher/__init__.py
@@ -1,5 +1,5 @@
-from PySide2.QtGui import *
-from PySide2.QtWidgets import *
+from PySide6.QtGui import *
+from PySide6.QtWidgets import *
 import webbrowser
 import os
 import sys
